FIX: Open block editor asynchronously after creation
