+++ apps/web/app/pages/sites/:id/editor/block/bar.tsx
@@ -1,3 +1,5 @@
+import { nanoid } from "nanoid";
+import { useState } from "react";
 import { css, styled } from "styled-components";
 import * as Typo from "@/components/ui/typography";
 import { rem } from "@/utils/rem";
@@ -11,16 +13,17 @@ import { Icon } from "@/components/ui/icon";
 import { useClientMutation } from "@/lib/react-query/useClientMutation";
 import { postSiteBlocks } from "@/apis/sites/pages/blocks";
 import { handleError } from "@/lib/axios/handleError";
-import { useSiteContext } from "../../contexts/site.context";
+import { useSiteContext } from "@/pages/sites/:id/contexts/site.context";
 import { getSiteBlockSlugValidation } from "@/apis/sites/pages/blocks/slug-duplication";
-import { nanoid } from "nanoid";
 import { BlockIcons, BlockNames } from "../constants";
+import { useAsyncOpenBlockEditor } from "../hooks/useAsyncOpenBlockEditor";
 
 export function BlockBar() {
   const { site, selectedPageId, refetchSite } = useSiteContext();
   const { mutateAsync: addBlock } = useClientMutation({
     mutationFn: postSiteBlocks,
   });
+  const [createdBlockId, setCreatedBlockId] = useState<number | null>(null);
 
   const handleClickBlockButton = async (data: CreateSiteBlockDto) => {
     const fn = async () => {
@@ -34,7 +37,7 @@ export function BlockBar() {
         },
       });
 
-      await addBlock({
+      const response = await addBlock({
         params: {
           siteId: site.id,
           pageId: selectedPageId,
@@ -46,7 +49,10 @@ export function BlockBar() {
             : data,
         },
       });
+
       await refetchSite();
+
+      setCreatedBlockId(response.data.block.id);
     };
 
     try {
@@ -61,6 +67,16 @@ export function BlockBar() {
       }
     }
   };
+
+  /**
+   * @useEffect
+   * Open created block in the editor after some amount of delay
+   * because the site data isn't updated right after the creation succeed
+   */
+  useAsyncOpenBlockEditor({
+    blockId: createdBlockId,
+    setBlockId: setCreatedBlockId,
+  });
 
   return (
     <BarBase>
