+++ apps/web/app/pages/sites/:id/editor/block/index.tsx
@@ -3,7 +3,6 @@ import { useEffect, useRef, useState, type MouseEvent } from "react";
 import { BlockRegistry } from "@clayout/kit";
 import { SiteBlockSchema, type SiteBlock } from "@clayout/interface";
 import {
-  useAddBlock,
   useBlockById,
   useRemoveBlock,
   useReorderBlock,
@@ -29,8 +28,8 @@ import { deleteSiteBlocks } from "@/apis/sites/pages/blocks";
 import { handleError } from "@/lib/axios/handleError";
 import { useDialog } from "@/components/ui/dialog";
 import { ConfirmDeleteDialog } from "@/components/shared/dialogs/confirm/delete";
-import { useTimer } from "@/hooks/useTimer";
 import { toast } from "sonner";
+import { useAsyncOpenBlockEditor } from "../hooks/useAsyncOpenBlockEditor";
 
 interface Props {
   blockId: string;
@@ -39,7 +38,6 @@ interface Props {
 
 export function Block({ blockId, blockIndex }: Props) {
   const blockRef = useRef<HTMLDivElement>(null);
-  const timer = useTimer();
   const [duplicatedBlockId, setDuplicatedBlockId] = useState<number | null>(
     null
   );
@@ -55,7 +53,6 @@ export function Block({ blockId, blockIndex }: Props) {
   const blockSchema = useBlockById(blockId);
   const reorderBlockLocally = useReorderBlock();
   const removeBlockLocally = useRemoveBlock();
-  const addBlockLocally = useAddBlock();
   const { mutateAsync: reorderBlock } = useClientMutation({
     mutationFn: postSiteBlockReorder,
   });
@@ -182,12 +179,6 @@ export function Block({ blockId, blockIndex }: Props) {
         },
       });
 
-      const parsed = SiteBlockSchema.safeParse(response.data.block);
-
-      if (parsed.success) {
-        addBlockLocally(selectedPage.id, parsed.data);
-      }
-
       await invalidateSiteCache();
 
       setDuplicatedBlockId(response.data.block.id);
@@ -270,26 +261,12 @@ export function Block({ blockId, blockIndex }: Props) {
   /**
    * @useEffect
    * Open duplicated block in the editor after some amount of delay
-   * because the site data isn't updated right after duplication success
+   * because the site data isn't updated right after the duplication succeed
    */
-  useEffect(() => {
-    if (duplicatedBlockId) {
-      if (timer.current) {
-        clearTimeout(timer.current);
-      }
-
-      timer.current = setTimeout(() => {
-        openBlockEditor(duplicatedBlockId);
-        setDuplicatedBlockId(null);
-      }, 100);
-    }
-
-    return () => {
-      if (timer.current) {
-        clearTimeout(timer.current);
-      }
-    };
-  }, [duplicatedBlockId, openBlockEditor, timer]);
+  useAsyncOpenBlockEditor({
+    blockId: duplicatedBlockId,
+    setBlockId: setDuplicatedBlockId,
+  });
 
   /**
    * @useEffect
